move avatarUpload to helpers
