+++ .env.example
@@ -2,3 +2,6 @@ PORT=
 JWT_SECRET=
 MONGO_URL=
 SENDGRID_API_KEY=
+CLOUD_NAME = 
+CLOUD_KEY =
+CLOUD_SECRET =
